+++ src/point_handling.c
@@ -6,7 +6,7 @@
 /*   By: bazaluga <bazaluga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/21 12:33:10 by bazaluga          #+#    #+#             */
-/*   Updated: 2024/05/29 13:08:45 by bazaluga         ###   ########.fr       */
+/*   Updated: 2024/06/14 15:14:00 by bazaluga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ t_point	*new_point(t_point **res, t_point p, t_fdata *data)
 t_point	*dyn_alloc_point_arr(t_pts_arr *pts, t_fdata *d)
 {
 	t_point	*arr;
-	int		i;
+	size_t	i;
 
 	if (pts->size == 0)
 		pts->size = SIZE_ARR;
